% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_classicOSLsignal.R
\name{calc_classicOSLsignal}
\alias{calc_classicOSLsignal}
\title{classic CW-OSL signal calculation}
\usage{
calc_classicOSLsignal(
  curve,
  components = NULL,
  algorithm = "late",
  signal.end.late = 0.8,
  signal.end.early = 0.4,
  stimulation.intensity = 35,
  verbose = TRUE
)
}
\value{
The input table \strong{components} \link{data.frame} will be returned with added/overwritten columns
}
\description{
Early background calculation after \link{1}, error calculation after (3) in \link{2}.
If the stimulation intensity differs from 35 mW /cm^-2 \link{1}, the intervals are increased/decreased accordingly
}
\details{
\link{1}A. C. Cunningham and J. Wallinga, ‘Selection of integration time intervals for quartz OSL decay curves’, Quaternary Geochronology, vol. 5, no. 6, pp. 657–666, Dec. 2010.
\link{2}R. F. Galbraith and R. G. Roberts, ‘Statistical aspects of equivalent dose and error calculation and display in OSL dating: An overview and some recommendations’, Quaternary Geochronology, vol. 11, pp. 1–27, Aug. 2012.
}
\section{Changelog}{

*
}

\section{ToDo}{

\itemize{
\item Update documentation
\item 
}
}

\section{Last changed}{
 2019-10-11
}

\examples{

}
\references{
Dirk Mittelstrass, Christoph Schmidt, Sebastian Kreutzer, ... .\emph{in preperation}. Algebraic CW-OSL
signal component decomposition of quartz and its use for natural dose evaluation
}
\seealso{
\link{decompose_OSLcurve}, \link{simulate_OSLcurve}
}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
\keyword{}
