% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_OSLintervals.R
\name{calc_OSLintervals}
\alias{calc_OSLintervals}
\title{Calculate integration intervals for CW-OSL deconvolution}
\usage{
calc_OSLintervals(
  components,
  curve = NULL,
  background.fitting = TRUE,
  channel.width = NA,
  channel.number = NA,
  t.start = 0,
  t.end = NA,
  verbose = TRUE
)
}
\arguments{
\item{components}{\link{data.frame} (\strong{required}):
Table containing the decay constants of the signal components}

\item{curve}{\link{data.frame} (\emph{optional}):
OSL signal curve. The x-axis (time axis) will be used ot define channel.width and channel.number}

\item{channel.width}{\link{numeric} (\emph{optional}):
channel width in seconds. Necessary if \emph{curve} is not given}

\item{channel.number}{\link{integer} (\emph{optional}):
number of channels resp. data points. Necessary if \emph{curve} is not given}

\item{t.start}{\link{numeric} (\emph{with default}):
starting point of the first interval, per default the start of the measurement}

\item{t.end}{\link{numeric} (\emph{optional}):
end point of the last interval, per default the end of the measurement}
}
\value{
The input table \emph{components} \link{data.frame} will be returned with four additional columns:
\emph{$t.start}, \emph{$t.end} defining the interval borders in time; \emph{$ch.start}, \emph{$ch.end} defining the intervals as channels
}
\description{
The function provides the integration intervals for CW-OSL component separation with \link{decompose_OSLcomponents}
}
\section{Changelog}{

\itemize{
\item 2018-04-05, DM: first running version
\item 2018-06-16, DM: added 2-component and 1-component case
\item 2018-06-19, DM: changed t0 and t3 parameter to t.start and t.end and added full increment proof
\item 2018-06-24, DM: changed data structure to get static tables for \link{analyse_SAR.CWOSL.deconvolved}
\item 2019-03-21, DM: Rewritten for arbitrary component numbers and changed data structure
\item 2019-10-02, DM: added Background-fitting interval determination
}
}

\section{ToDo}{

\itemize{
\item optional data.frame with result-matrix
\item optional RLum.Data.Curve as input and error calculation as result-matrix
\item more comments
}
}

\section{Last changed}{
 2019-10-02
}

\examples{
...

}
\references{
Dirk Mittelstrass, Christoph Schmidt, Sebastian Kreutzer, ... .\emph{in preperation}. Algebraic CW-OSL
signal component decomposition of quartz and its use for dose evaluation within the R luminescence package
}
\seealso{
\link{decompose_OSLcomponents}, \link{...}, \link{..}
}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
\keyword{CWOSL}
\keyword{OSL}
\keyword{components}
\keyword{decomposition}
\keyword{deconvolution}
