% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_RLum.Data.R
\name{check_RLum.Data}
\alias{check_RLum.Data}
\title{Check if an object is a valid RLum.Data.Curve record for use in RLum.OSL functions}
\usage{
check_RLum.Data(
  object,
  record_type = "OSL",
  curve_template = NULL,
  verbose = TRUE
)
}
\arguments{
\item{object}{\link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} (\strong{required}):
Input object which shall be tested.}

\item{record_type}{\link{character} (\emph{with default}):
Expected type of record of the input \code{object},
for example: \code{"OSL"},\code{"SGOSL"} or \code{"IRSL"}.}

\item{curve_template}{\link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} (\emph{optional}):
Curve to check x-axis and some measurement parameter against.}

\item{verbose}{\link{logical} (\emph{with default}):
Enables console text output.}
}
\value{
A bolean value: \code{TRUE} or \code{FALSE}.
}
\description{
The input object is checked for the following properties:
\itemize{
\item Is the object of class \link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve} ?
\item Has the object a slot \code{recordType}?
\item Does the objects record type match with this functions argument \code{record_type}?
\item Is the record not just a XSYG metadata object (marked by '_' before the record type name)?
\item Contains the record a curve, thus has the object a slot \code{data}?
\item Is the curve of type XY, thus has it a 2 x N dimension?
\item If a \code{curve_template} is provided, the input object is also checked if it matches number of data points, x-axis and the \code{info} parameters "LPOWER", "LTYPE" and "TEMPERATURE".
}

If all checks are positive, the input object is regarded as suitable for the functions
\link{RLum.OSL_correction}, \link{RLum.OSL_global_fitting}, \link{RLum.OSL_decomposition} and other functions if their
input curve is of type \link[Luminescence:RLum.Data.Curve-class]{Luminescence::RLum.Data.Curve}.
}
\section{Last updates}{


2026-02-17, DM: Created function.
}

\examples{

# Load example data
data_path <- system.file("examples", "FB_10Gy_SAR.bin", package = "OSLdecomposition")
data_set <- Luminescence::read_BIN2R(data_path, fastForward = TRUE)

# Test if record is of type OSL
check_RLum.Data(data_set[[5]]@records[[1]])

}
\author{
Dirk Mittelstraß, \email{dirk.mittelstrass@luminescence.de}
}
