% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_SARdata.R
\name{decompose_SARdata}
\alias{decompose_SARdata}
\title{Analyse SAR CW-OSL measurements using OSL decomposition}
\usage{
decompose_SARdata(
  object,
  aliquot_selection = NULL,
  components,
  record_type = "OSL",
  algorithm = "det+nls",
  recuperation_rate = 0.05,
  recycling_ratio = 0.1,
  dose.rate = 1,
  dose.points = NULL,
  applied.pre.dose = 0,
  fit.method = "EXP",
  negative_values_to_zero = TRUE,
  data_has_Tx = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{\linkS4class{RLum.Analysis} or \href{RLum.Analysis}{list} (\strong{required}):
data set of one or multiple aliquots containing containing SAR CW-OSL measurements}

\item{aliquot_selection}{\link{c} (\emph{optional}):
vector specifying which items (aliquots) of a list of \linkS4class{RLum.Analysis} objects shall be combined}

\item{components}{\link{data.frame} (\strong{required}):
Table containing the decay parameters of the OSL curve. One column must be named \emph{$lambda}.
It is recommended to provide also the integration interval parameters (columns \emph{t.start, t.end, ch.start, ch.end}),
which can be found by applying \link{calc_OSLintervals} on the global mean curve, calculated by \link{sum_OSLcurves}.
If one or more column is missing, \link{calc_OSLintervals} is run automatically.}

\item{record_type}{\link{character} (\emph{with default}):
record type ("OSL","SGOSL","IRSL") which shall be evaluated}

\item{recuperation_rate}{\link{numeric} (\emph{with default}):
check aliquot data for the recuperation rate as rejection criteria; set to 'NA' for ignoring}

\item{recycling_ratio}{\link{numeric} (\emph{with default}):
check aliquot data for the recycling ratio as rejection criteria; set to 'NA' for ignoring}

\item{dose.points}{\link{c} (\emph{optional}):
a numeric vector containg the dose points values Using this argument
overwrites dose point values in the signal curves.}

\item{negative_values_to_zero}{\link{logical} (\emph{with default}): if TRUE negative n values are overwritten and set to n = 0;
negative n values are mathematically correct but physically debatable and can cause problems with \link[Luminescence:plot_GrowthCurve]{Luminescence::plot_GrowthCurve}}

\item{data_has_Tx}{\link{logical} (\emph{with default}):
if TRUE it is assumed that the sequence have a test dose steps to normalize the regeneration dose points.
If FALSE, every dose point is used as regeneration dose point}
}
\value{
A \link{list} is returned containing the following elements:

\item{[\link{i}]}{\strong{list entry} RLum object related to the fast OSL decay component}
\item{...$data}{\link{data.frame} contains De results from \link{plot_GrowthCurve}}
\item{...$LnLxTnTx.table}{\href{[data.frame]}{list} LxTx tables for each \linkS4class{RLum.Analysis} object (aliquot)}
\item{...$rejection.criteria}{\href{[data.frame]}{list} rejection criteria tables for each \linkS4class{RLum.Analysis} object (aliquot)}
\item{...$Formula}{\href{[character]}{list} growth curve fit formula for each \linkS4class{RLum.Analysis} object (aliquot)}
}
\description{
Calculates separate growth curves for each OSL signal component.
Alternative for \link[Luminescence:analyse_SAR.CWOSL]{Luminescence::analyse_SAR.CWOSL}
}
\section{Changelog}{

\itemize{
\item 2018-05-18, DM: first running version
\item 2018-05-27, DM: first version with good chance of suceeding (for Bayreuth meeting 2018-05-28)
\item 2018-07-02, DM: try(plot_GrowthCurve(...)); more integrity checks for rejection critera calculation
\item 2019-04-30, DM: deleted fit-parameter-evaluation stuff; works now for any number of components; renamed function
\item 2019-10-11, DM: Tweaked for calc_classicOSLsignals() and Rmarkdown-script
}
}

\section{ToDo}{

\itemize{
\item streamline code
\item add capability to include background fitting
}
}

\section{Last changed}{
 2019-10-27
}

\examples{

}
\references{
Dirk Mittelstrass, Christoph Schmidt, Sebastian Kreutzer, ... .\emph{in preperation}. Algebraic CW-OSL
signal component decomposition of quartz and its use for dose evaluation within the R luminescence package
}
\seealso{
\link{...}, \link{...}, \link{...}
}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
\keyword{CW-OSL}
\keyword{CWOSL}
\keyword{OSL}
\keyword{SAR}
\keyword{components}
\keyword{decompostion}
