% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2020-05-07_fit_OSLcurve.R, R/fit_OSLcurve.R
\name{fit_OSLcurve}
\alias{fit_OSLcurve}
\title{Fit global mean curve to get decay constants}
\usage{
fit_OSLcurve(
  curve,
  K.max = 3,
  F.threshold = 150,
  stimulation.intensity = 30,
  stimulation.wavelength = 470,
  algorithm = "DE+LM",
  parallel.computing = TRUE,
  verbose = TRUE,
  output.complex = TRUE
)

fit_OSLcurve(
  curve,
  K.max = 3,
  F.threshold = 150,
  stimulation.intensity = 30,
  stimulation.wavelength = 470,
  algorithm = "DE+LM",
  parallel.computing = TRUE,
  verbose = TRUE,
  output.complex = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{



}
\description{
Fit global mean curve to get decay constants

Fit global mean curve to get decay constants
}
\section{Changelog}{

* 2019-02-14 First version
* 2019-03-15 Seperated 'decompose_OSLalternatively()'
* 2019-04-29 Added Blucszs & Adamiec-like approach using numOSL::decomp (Peng et al. 2014)
* 2019-05-14 Added "fit_OSLLifeTimes" approach from Luminescence package 0.9; Corrected and improved numOSL approach; Deleted nls.default approach
* 2019-06-28 Deleted "fit_OSLLifeTimes" approach. Added stretched exponentials for testing. Added overview plot
* 2019-10-07 Streamlined function; added optional background fitting
* 2019-10-08 Seperated plotting to plot_PhotoCrosssections()
* 2020-04-04 Extended output list (curve & arguments)
* 2020-04-06 Extended print output and made some  tweaks. Replaced 'SAR.compatible' with 'fully.bleached'
* 2020-05-05 Replaced 'fully.bleached' with 'bleaching.grade'


* 2019-02-14, DM: First version
* 2019-03-15, DM: Seperated 'decompose_OSLalternatively()'
* 2019-04-29, DM: Added Blucszs & Adamiec-like approach using numOSL::decomp (Peng et al. 2014)
* 2019-05-14, DM: Added "fit_OSLLifeTimes" approach from Luminescence package 0.9; Corrected and improved numOSL approach; Deleted nls.default approach
* 2019-06-28, DM: Deleted "fit_OSLLifeTimes" approach. Added stretched exponentials for testing. Added overview plot
* 2019-10-07, DM: Streamlined function; added optional background fitting
* 2019-10-08, DM: Seperated plotting to plot_PhotoCrosssections()
* 2020-04-04, DM: Extended output list (curve & arguments)
* 2020-04-06, DM: Extended print output and made some  tweaks. Replaced 'SAR.compatible' with 'fully.bleached'
* 2020-05-05, DM: Replaced bolean 'fully.bleached' with numeric 'bleaching.grade'
* 2020-08-05, DM: Added DEoptim + nlsLM algorithm
* 2020-08-10, DM: Optional parallel computing enabled
}

\section{ToDo}{

* ! Replace numOSL::decomp by own code. Then enable sigma.lambda calculation and replace "easy way" of intensity calculation !
* Write documentation
* Reconsider background value fitting


* Write documentation
* Reactivate optional background level fitting
* Enable list of DE and LM control parameters as argument
* Enable optional weighted fitting and give out reduced Chi²
}

\section{Last changed. 2020-05-07}{
NA
}

\section{Last changed. 2020-08-12}{
NA
}

\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}

Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
