% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimise_OSLintervals.R
\name{optimise_OSLintervals}
\alias{optimise_OSLintervals}
\title{Find adequate integration intervals for CW-OSL decomposition}
\usage{
optimise_OSLintervals(
  components,
  curve = NULL,
  background.fitting = FALSE,
  channel.width = NA,
  channel.number = NA,
  t.start = 0,
  t.end = NA,
  verbose = TRUE
)
}
\arguments{
\item{components}{\link{data.frame} (\strong{required}):
Table containing the decay constants of the signal components}

\item{curve}{\link{data.frame} (\emph{optional}):
OSL signal curve. The x-axis (time axis) will be used ot define channel.width and channel.number}

\item{channel.width}{\link{numeric} (\emph{optional}):
channel width in seconds. Necessary if \emph{curve} is not given}

\item{channel.number}{\link{integer} (\emph{optional}):
number of channels resp. data points. Necessary if \emph{curve} is not given}

\item{t.start}{\link{numeric} (\emph{with default}):
starting point of the first interval, per default the start of the measurement}

\item{t.end}{\link{numeric} (\emph{optional}):
end point of the last interval, per default the end of the measurement}
}
\value{
The input table \emph{components} \link{data.frame} will be returned with four additional columns:
\emph{$t.start}, \emph{$t.end} defining the interval borders in time; \emph{$ch.start}, \emph{$ch.end} defining the intervals as channels
}
\description{
The function optimises the integration intervals for CW-OSL component separation with \link{decompose_OSLcurve}.
}
\section{Changelog}{

\itemize{
\item 2018-04-05, DM: first running version
\item 2018-06-16, DM: added 2-component and 1-component case
\item 2018-06-19, DM: changed t0 and t3 parameter to t.start and t.end and added full increment proof
\item 2018-06-24, DM: changed data structure to get static tables
\item 2019-03-21, DM: Rewritten for arbitrary component numbers and changed data structure again
\item 2019-10-02, DM: added optional determination of background-fitting interval
\item 2020-08-23, DM: Replaced own minimum searching algorithm with DEoptim
\item 2020-08-27, DM: Renamed function; More input classes allowed
}
}

\section{ToDo}{

\itemize{
\item rename function into "optimise_OSLintervals" and edit depending functions
\item write full documentation
}
}

\section{Last changed}{
 2020-08-27
}

\examples{

A <- optimise_OSLintervals(c(2, 0.5, 0.02), channel.width = 0.1, channel.number = 200, background.fitting = TRUE)
print(A, row.names = FALSE)

}
\references{
Mittelstraß, D., Schmidt, C., Beyer, J., Heitmann, J. and Straessner, A.:
Automated identification and separation of quartz CW-OSL signal components with R, \emph{in preparation}.
}
\seealso{
\link{decompose_OSLcomponents}, \link{RLum.OSL_decomposition}, \link[DEoptim:DEoptim]{DEoptim::DEoptim}
}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
\keyword{CWOSL}
\keyword{OSL}
\keyword{components}
\keyword{decomposition}
\keyword{deconvolution}
