% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OSLcurve.R
\name{plot_OSLcurve}
\alias{plot_OSLcurve}
\title{Advanced plot function for CW-OSL curves and CW-OSL signal components}
\usage{
plot_OSLcurve(
  curve = NULL,
  components,
  display = "detailed",
  show.intervals = FALSE,
  title = NULL,
  hide.plot = FALSE,
  filename = NULL
)
}
\arguments{
\item{display}{Choices: "detailed" (default), "components", "raw" ,"compare_lin", "compare_log", "lin-log"}

\item{show.intervals}{Draws vertical lines into the residual plot showing the signal bin intervals for Step 2 CW-OSL decomposition}

\item{title}{Plot title}

\item{hide.plot}{If true, plot is not drawn but can be catched by 'A <- plot_OSLcurve(...)' or saved as file.
If catched, the plot ca be displayed later for example by [gridExtra::grid.arrange]}

\item{filename}{Filename or path to save the diagram as file. If just a name is chosen, the file is
saved in the working directory. The graphic type is chosen according to the file ending.
Allowed are .pdf, .eps, .svg (vector graphics) and .jpg, .png, .bmp (pixel graphics)
and more, see [ggplot2::ggsave]}
}
\description{
Advanced plot function for CW-OSL curves and CW-OSL signal components
}
\section{Notes}{


pseudoLM-OSL curves are created using the transformation described by Bulur (2000).
The stimulation ramp duration is chosen double the CW-OSL duration.
See also Bos and Wallinga (2012) for a detailed explanation and discussion
}

\section{Last changed. 2020-09-13}{
NA
}

\examples{

# Set some reasonable parameter for a weak quartz CW-OSL decay
components <- data.frame(name = c("fast", "medium", "slow"), lambda = c(2, 0.5, 0.02), n = c(1000, 1000, 10000))

# Simulate the CW-OSL curve and add some signal noise
curve <- simulate_OSLcurve(components, simulate.curve = TRUE, add.poisson.noise = TRUE)

# Display the simulated curve
plot_OSLcurve(curve, components)

Changelog:
* 2019-03-06, DM: First reasonable version
* 2019-04-03, DM: Rebuild whole function
* 2019-10-02, DM: Added background component support and some little tweaks
* 2020-04-22, DM: Enabled hidden output to draw later
* 2020-06-19, DM: Added pseudoLM-OSL diagrams
* 2020-08-04, DM: Added subtitles and RSS info
* 2020-09-02, DM: Added graphic saving with [ggplot2::ggsave]

ToDo:
* REFACTORIZE CODE
* ! Put the ggplot building (or at least its style options) in its own sub-function !
* Add input checks and data conversions from [decompose_OSLcurve]
* When drawing components without curve, skip residual curve
* Display fitting formula
* Get rid of library 'ggpubr' to decrease dependencies
* Change from library("XXX") to XXX::
* Add argument ggsave.control() to give direct control about image saving
* Add table column for the lambda error
}
\references{
Bos, A. J. J. and Wallinga, J.: How to visualize quartz OSL signal components, Radiation Measurements, 47(9), 752–758, doi:10.1016/j.radmeas.2012.01.013, 2012.

Bulur, E.: A simple transformation for converting CW-OSL curves to LM-OSL curves, Radiation Measurements, 32(2), 141–145, doi:10.1016/S1350-4487(99)00247-4, 2000.

Bos, A. J. J. and Wallinga, J.: How to visualize quartz OSL signal components,
Radiation Measurements, 47(9), 752–758, doi:10.1016/j.radmeas.2012.01.013, 2012.

Bulur, E.: A simple transformation for converting CW-OSL curves to LM-OSL curves,
Radiation Measurements, 32(2), 141–145, doi:10.1016/S1350-4487(99)00247-4, 2000.

Mittelstraß, D., Schmidt, C., Beyer, J., Heitmann, J. and Straessner, A.:
Automated identification and separation of quartz CW-OSL signal components with R, *in preparation*.
}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
