% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OSLresults.R
\name{plot_OSLresults}
\alias{plot_OSLresults}
\title{Advanced plot function for OSL decomposition results}
\usage{
plot_OSLresults(
  data,
  parameter = "De",
  max = NA,
  data2 = NULL,
  parameter2 = "De",
  max2 = NA,
  correlate = FALSE,
  title = NULL
)
}
\description{
Advanced plot function for OSL decomposition results
}
\section{Changelog}{

* 2019-10-28, DM: First reasonable version
}

\section{ToDo}{

* #'
}

\section{Last changed. 2019-10-28}{
NA
}

\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
