% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PhotoCrosssections.R
\name{plot_PhotoCrosssections}
\alias{plot_PhotoCrosssections}
\title{Plot comparison of the global OSL curve photo-crosssections with literature values}
\usage{
plot_PhotoCrosssections(
  fit.result.list,
  stimulation.intensity = 35,
  stimulation.wavelength = 470,
  title = NULL
)
}
\arguments{
\item{fit.result.list}{}
}
\value{

}
\description{
Plot comparison of the global OSL curve photo-crosssections with literature values
}
\section{Changelog}{

* 2019-10-08 Seperated from fit_OSLcurve()
}

\section{ToDo}{

* Clean code
* Documentation
* Sometimes the red rectangle is not drawn (fit_OSLcurve(Batagai, K.max = 6)). Why?
* Add Slow1, Slow2 etc into output.plot
* for other than just 470nm values
}

\section{Last changed. 2019-10-25}{
NA
}

\examples{

}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
