% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_OSLdata.R
\name{prepare_OSLdata}
\alias{prepare_OSLdata}
\title{CW-OSL preperation}
\usage{
prepare_OSLdata(
  object,
  record.type = "OSL",
  cut.time = NA,
  background.aliquot = NA,
  background.plot = TRUE,
  tailor.single.grain = FALSE
)
}
\description{
The function prepares raw CW-OSL data sets for OSL decomposition process.
}
\section{Changelog}{

}

\section{ToDo}{

\itemize{
\item Give back data set structure
\item Correct x-axis, if it begins with Zero
\item Delete preset aliquots
\item Get vector with OSL record indicies
\item test sequence on SAR compability for each aliquot
}
}

\section{Last changed}{
 2019-10-25
}

\examples{

}
\references{
Dirk Mittelstrass, Christoph Schmidt, Sebastian Kreutzer, ... .\emph{in preperation}. Algebraic CW-OSL
signal component decomposition of quartz and its use for natural dose evaluation
}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
\keyword{CW-OSL}
\keyword{CWOSL}
\keyword{OSL}
\keyword{component}
\keyword{components}
\keyword{decomposition}
\keyword{deconvolution}
