% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.OSL_correction.R
\name{RLum.OSL_correction}
\alias{RLum.OSL_correction}
\title{Title}
\usage{
RLum.OSL_correction(
  object,
  record_type = "OSL",
  background_sequence = NULL,
  offset = 0,
  check_consistency = TRUE,
  check_signal_level = FALSE,
  tailor_records = FALSE,
  record_cut = NA,
  verbose = TRUE,
  report = FALSE
)
}
\arguments{
\item{object}{[RLum.Analysis-class] (**required**):
input object containing data for analysis, alternatively a [list] of
[RLum.Analysis-class] objects can be provided.}

\item{tailor_records}{**Inactive**. This algorithm cut}

\item{record_cut}{}

\item{report}{}
}
\value{

}
\description{
Title
}
\examples{
library(Luminescence)

}
