% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.OSL_decomposition.R
\name{RLum.OSL_decomposition}
\alias{RLum.OSL_decomposition}
\title{Decomposes RLum.Data.Curves into its CW-OSL components}
\usage{
RLum.OSL_decomposition(
  object,
  record_type = "OSL",
  decay_rates = NULL,
  error_calculation = "empiric",
  report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{decay_rates}{If not defined, the decay rates are taken from object$OSL_decomponents$case.tables[[object$OSL_decomponents$K.selected]].
Therefore, the number of components can be changed by altering object$OSL_decomponents$K.selected}

\item{verbose}{}
}
\value{

}
\description{
Decomposes RLum.Data.Curves into its CW-OSL components
}
