% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_SARdata.R
\name{table_SARdata}
\alias{table_SARdata}
\title{Creates result table of the SAR analysis}
\usage{
table_SARdata(
  SAR.list,
  data.type = "De",
  criterion = "median",
  apply.RC = FALSE,
  age_rate = NA,
  unit = "",
  digits = 3,
  add.number = FALSE,
  add.number.prefix = "(",
  expected_value = NA,
  output.plot = FALSE,
  title = NULL
)
}
\description{
Creates result table of the SAR analysis
}
\section{Changelog}{

* 2019-10-14, DM: first version
}

\section{ToDo}{

* nice output with ggtexttable or formattable or kable
* refactorize, streamline and comment whole funciton
* Roxygen documentation
}

\section{Last changed}{
 2019-10-24
}

\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
