% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_OSLcurve.R
\name{decompose_OSLcurve}
\alias{decompose_OSLcurve}
\title{CW-OSL decomposition}
\usage{
decompose_OSLcurve(
  curve,
  components,
  background.fitting = FALSE,
  algorithm = "det+nls",
  error.calculation = "empiric",
  verbose = TRUE
)
}
\arguments{
\item{curve}{\linkS4class{RLum.Data.Curve} or \link{data.frame} (\strong{required}):
CW-OSL record. First column (x-axis) must contain time marks, Second column (y-axis) must contain signal values. Further columns will be ignored}

\item{components}{\link{data.frame} (\strong{required}):
Template table containing the decay parameters of the OSL curve. One column must be named \emph{$lambda}.
It is recommended to provide also the integration interval parameters (columns \emph{t.start, t.end, ch.start, ch.end}),
which can be found by applying \link{calc_OSLintervals} on the global mean curve, calculated by \link{sum_OSLcurves}.
If one or more column is missing, \link{calc_OSLintervals} is run automatically.}

\item{error.calculation}{(\emph{with default}):
integral error estimation approach, either \strong{"empiric"} or \strong{"poisson"} or *\emph{numerical value};
Per default the data of \emph{curve$residual} provided by \link{simulate_OSLcurve} is used to calculate an \strong{empiric} standard error for each integral which will be processed in the error propagation formula.
Alternatively the integral standard error can be calculated by assuming a \strong{poisson} distributed signal error, known as \emph{Shot noise}.
This is suitable if the lack of data points on the x-axis circumvent an empiric error estimation, like with spatial or spectral resolved CCD measurements.
Also the parameter can be set to a \strong{numerical value} which will be handled as standard deviation per channel and added to the poisson distributed \emph{Shot noise}}
}
\value{
The input table \strong{components} \link{data.frame} will be returned with added/overwritten columns: \emph{$n, $n.error, $n.residual, $I, $I.error}
}
\description{
The function calculates the CW-OSL component amplitudes by a determinant-based algorithm.
It also estimates the standard deviation of the amplitudes by using the error propagation method.
}
\section{Changelog}{

\itemize{
\item winter 2012/13: first version, used for Bachelor thesis DM
\item autumn 2013   : added empiric error estimation, shown in germanLED Freiberg 2013
\item 2014-11-??, SK: formated into Rluminecence package standard
\item 2014-11-07, DM: Binomial error propagation added
\item 2018-05-01, DM: added interval time arguments and many little tweaks
\item 2018-05-04, DM: added residuals for n values (necessary for slow component dosimetry)
\item 2018-06-22, DM: added deconvolution of data sets with 1 or 2 components
\item 2018-06-22, DM: added negative.values.to.zero and set it on TRUE as default (on request of C. Schmidt)
\item 2018-07-05, DM: overworked error estimation; replaced binomial with poisson error approach; added auto-switch to poisson if integral length = 1; integrated background.noise into error
\item 2019-03-29, DM: Rewritten function for several purposes: 1. working now with any number of components  2. shorter and more elegant code 3. data format more suitable for markdown/shiny applications.
\item 2019-09-09, DM: Added anticorrelating covariance terms into error estimation
\item 2019-09-25, DM: Merged function with decompose_OSLalternatively() and added algorithm argument
\item 2019-09-25, DM: Deleted unnecessary stuff (negative.values.to.zero, offset, anticorrelation)
\item 2019-10-02, DM: Added background fitting
}
}

\section{ToDo}{

\itemize{
\item Update documentation
\item Print running time info
\item Print output table
\item Add measurement-cut-parameter. Set $remainder of Slow components to NA if used
}
}

\section{Last changed}{
 2019-10-07
}

\examples{

}
\references{
Dirk Mittelstrass, Christoph Schmidt, Sebastian Kreutzer, ... .\emph{in preperation}. Algebraic CW-OSL
signal component decomposition of quartz and its use for natural dose evaluation
}
\seealso{
\link{calc_OSLintervals}, \link{simulate_OSLcurve}, \link{sum_OSLcurves}
}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
\keyword{CW-OSL}
\keyword{CWOSL}
\keyword{OSL}
\keyword{component}
\keyword{components}
\keyword{decomposition}
\keyword{deconvolution}
