% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_OSLcurve.R
\name{decompose_OSLcurve}
\alias{decompose_OSLcurve}
\title{CW-OSL decomposition}
\usage{
decompose_OSLcurve(
  curve,
  components,
  background.fitting = FALSE,
  algorithm = "det",
  error.calculation = "empiric",
  verbose = TRUE
)
}
\arguments{
\item{curve}{\linkS4class{RLum.Data.Curve} or \link{data.frame} (\strong{required}):
CW-OSL record. \verb{$time} or first column (x-axis) for the measurement time (must have constand time intervals),
\verb{$signal} or second column (y-axis) for the signal values. Further columns will be ignored}

\item{components}{\link{data.frame} or \link{numeric} vector (\strong{required}):
Either a vector containing the decay parameters of the CW-OSL components or a table (data.frame), usually the table returned by \link{fit_OSLcurve}.
In case of a vector: It is recommended to use less than 7 parameters. The parameters will be sorted in decreasing order.
In case of a data.frame. One column must be named \emph{$lambda}.
It is recommended to provide also integration interval parameters (columns \emph{t.start, t.end, ch.start, ch.end}),
which can be found by applying \link{calc_OSLintervals} on the global mean curve, calculated by \link{sum_OSLcurves}.
If one or more column is missing, a simple interval definition algorithm is run automatically (see \strong{Notes})..}

\item{error.calculation}{(\emph{with default}):
integral error estimation approach, either \strong{"empiric"} or \strong{"poisson"} or \link{numeric} value;
Per default the data of \emph{curve$residual} provided by \link{simulate_OSLcurve} is used to calculate an \strong{empiric} standard error for each integral which will be processed in the error propagation formula.
Alternatively the integral standard error can be calculated by assuming a \strong{poisson} distributed signal error, known as \emph{Shot noise}.
This is suitable if the lack of data points on the x-axis circumvent an empiric error estimation, like with spatial or spectral resolved CCD measurements.
Also the parameter can be set to a \link{numeric} value; which will be handled as standard deviation per channel and added to the poisson distributed \emph{Shot noise}.
If no error calculation is wished, set the argument to \strong{"none"}. This reduces the necessary computing time heavily.}
}
\value{
The input table \strong{components} \link{data.frame} will be returned with added/overwritten columns: \emph{$n, $n.error, $n.residual, $I, $I.error}
}
\description{
The function calculates the CW-OSL component amplitudes by a determinant-based algorithm.
It also estimates the standard deviation of the amplitudes by using the error propagation method.
}
\section{Changelog}{

\itemize{
\item winter 2012/13: first version, used for Bachelor thesis DM
\item autumn 2013   : added empiric error estimation, shown in germanLED Freiberg 2013
\item 2014-11-??, SK: formated into Rluminecence package standard
\item 2014-11-07, DM: Binomial error propagation added
\item 2018-05-01, DM: added interval time arguments and many little tweaks
\item 2018-05-04, DM: added residuals for n values (necessary for slow component dosimetry)
\item 2018-06-22, DM: added deconvolution of data sets with 1 or 2 components
\item 2018-06-22, DM: added negative.values.to.zero and set it on TRUE as default (on request of C. Schmidt)
\item 2018-07-05, DM: overworked error estimation; replaced binomial with poisson error approach; added auto-switch to poisson if integral length = 1; integrated background.noise into error
\item 2019-03-29, DM: Rewritten function for several purposes: 1. working now with any number of components  2. shorter and more elegant code 3. data format more suitable for markdown/shiny applications.
\item 2019-09-09, DM: Added anticorrelating covariance terms into error estimation
\item 2019-09-25, DM: Merged function with decompose_OSLalternatively() and added algorithm argument
\item 2019-09-25, DM: Deleted unnecessary stuff (negative.values.to.zero, offset, anticorrelation)
\item 2019-10-02, DM: Added optional background fitting
\item 2020-04-06, DM: Added 'initial.signal' column in output data.frame; cleaned print output
\item 2020-07-20, DM: Added algorithm for fast interval definition based on logarithmic means; More input data checks
}
}

\section{ToDo}{

\itemize{
\item Update documentation (example, notes)
\item Substitute Cramers rule determinant equation with solve() to increase performance
\item In some very rare cases, negative value for n.error are returned. Why?
}
}

\section{Last changed}{
 2020-07-20
}

\examples{

}
\references{

}
\seealso{
\link{calc_OSLintervals}, \link{fit_OSLcurve}, \link{sum_OSLcurves}
}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
\keyword{CW-OSL}
\keyword{CWOSL}
\keyword{OSL}
\keyword{component}
\keyword{components}
\keyword{decomposition}
\keyword{deconvolution}
