% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_OSLcurve.R
\name{fit_OSLcurve}
\alias{fit_OSLcurve}
\title{multi-exponential CW-OSL curve fitting}
\usage{
fit_OSLcurve(
  curve,
  K.max = 5,
  F.threshold = 150,
  stimulation.intensity = 30,
  stimulation.wavelength = 470,
  verbose = TRUE,
  output.complex = FALSE,
  parallel.computing = FALSE
)
}
\arguments{
\item{output.complex}{}
}
\value{

}
\description{
Fitting function for measurements of multi-exponential decay signal with first order kinetics.
This function is based on Bluszcz & Adamiec (2006) and was primarily developed for the analysis of quartz CW-OSL
measurements for dating and dosimetry applications.
}
\section{Changelog}{

* 2019-02-14, DM: First version
* 2019-03-15, DM: Seperated 'decompose_OSLalternatively()'
* 2019-04-29, DM: Added Blucszs & Adamiec-like approach using numOSL::decomp (Peng et al. 2014)
* 2019-05-14, DM: Added "fit_OSLLifeTimes" approach from Luminescence package 0.9; Corrected and improved numOSL approach; Deleted nls.default approach
* 2019-06-28, DM: Deleted "fit_OSLLifeTimes" approach. Added stretched exponentials for testing. Added overview plot
* 2019-10-07, DM: Streamlined function; added optional background fitting
* 2019-10-08, DM: Seperated plotting to plot_PhotoCrosssections()
* 2020-04-04, DM: Extended output list (curve & arguments)
* 2020-04-06, DM: Extended print output and made some  tweaks. Replaced 'SAR.compatible' with 'fully.bleached'
* 2020-05-05, DM: Replaced bolean 'fully.bleached' with numeric 'bleaching.grade'
* 2020-08-05, DM: Added DEoptim + nlsLM algorithm
* 2020-08-10, DM: Optional parallel computing enabled
}

\section{ToDo}{

* Complete documentation
* Reactivate optional background level fitting
* Introduce 'fit_OSLcurve.control' which forwards algorith parameters to DEoptim.control and nls.lm.control
* Enable optional weighted fitting and give out reduced Chi²
}

\section{Last changed}{
 2020-08-18
}

\examples{

# Create curve with two components
curve <- cbind(X = c(1, 2, 3, 4, 5, 6, 7, 8, 9), Y = c(42, 20, 12, 7, 5, 4, 3.1, 2.4, 2))

# Perform fitting
components <- fit_OSLcurve(curve, F.threshold = 3)

# Display results
plot_OSLcurve(curve, components)

}
\references{
Mittelstraß, D., Schmidt, C., Beyer, J., Heitmann, J. and Straessner, A.:
Automated identification and separation of quartz CW-OSL signal components with R, *in preparation*.
}
\seealso{
[RLum.OSL_decomposition], [sum_OSLcurves], [decompose_OSLcurve], [plot_OSLcurve],
[plot_PhotoCrosssections], [minpack.lm::nlsLM], [DEoptim::DEoptim]
}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
