% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_OSLcurve.R
\name{fit_OSLcurve}
\alias{fit_OSLcurve}
\title{Fit global mean curve to get decay constants}
\usage{
fit_OSLcurve(
  curve,
  K.max = 5,
  F.threshold = 50,
  stimulation.intensity = 35,
  stimulation.wavelength = 470,
  applied.time.cut = FALSE,
  weight.Chi = FALSE,
  background.fitting = FALSE,
  verbose = FALSE,
  output.complex = TRUE,
  output.plot = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Fit global mean curve to get decay constants
}
\section{Changelog}{

* 2019-02-14 First version
* 2019-03-15 Seperated 'decompose_OSLalternatively()'
* 2019-04-29 Added Blucszs & Adamiec-like approach using numOSL::decomp (Peng et al. 2014)
* 2019-05-14 Added "fit_OSLLifeTimes" approach from Luminescence package 0.9; Corrected and improved numOSL approach; Deleted nls.default approach
* 2019-06-28 Deleted "fit_OSLLifeTimes" approach. Added stretched exponentials for testing. Added overview plot
* 2019-10-07 Streamlined function; added optional background fitting
* 2019-10-08 Seperated plotting to plot_PhotoCrosssections()
}

\section{ToDo}{

* Documentation
* Test background value fitting
* Seperate crosssection calculation
}

\section{Last changed. 2019-10-19}{
NA
}

\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
