% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSLdecomposition-package.R
\docType{package}
\name{OSLdecomposition-package}
\alias{OSLdecomposition-package}
\alias{OSLdecomposition}
\title{Identify and Separate Continuous-Wave Optically Stimulated Luminescence Signal Components\cr

\if{html}{
\figure{OSLdecomposition_logo.png}{options: width="50px" alt="luminescence.de"}\cr
}}
\description{
Function library for the identification and separation of exponentially
decaying signal components in continuous-wave optically stimulated luminescence (CW-OSL) measurements.
A special emphasis is laid on luminescence dating with quartz, which is known for
systematic errors due to CW-OSL signal components with unequal physical behaviour.
This package enables an automatised workflow for signal component deconvolution
in combination with the R package 'Luminescence'.
}
\details{
\strong{Project website}
\itemize{
\item \url{https://luminescence.de}
}

\strong{Source code repository}
\itemize{
\item \url{https://github.com/DirkMittelstrass/OSLdecomposition}
}

\strong{Support and bug reporting}
\itemize{
\item \email{dirk.mittelstrass@luminescence.de} or
\item \url{https://github.com/DirkMittelstrass/OSLdecomposition/issues}
}

\strong{Related projects}
\itemize{
\item \url{https://r-luminescence.org}
}

\strong{Package maintainer}

Dirk Mittelstraß, Independent researcher, Dresden (Germany),\cr
\email{dirk.mittelstrass@luminescence.de}

\strong{Funding}
Dirk Mittelstraß created this package as part of his master thesis and further enhanced and
published it as private endeavour. He did not receive any specific grant from funding agencies
in the public, commercial or not-for-profit sectors.
}
\keyword{package}
