% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OSLdecomposition-package.R
\docType{package}
\name{OSLdecomposition-package}
\alias{OSLdecomposition-package}
\alias{OSLdecomposition}
\title{Identify and Separate Continuous-Wave Optically Stimulated Luminescence Signal Components}
\description{
Function library for the identification and separation of exponentially
decaying signal components in continuous-wave optically stimulated luminescence (CW-OSL) measurements.
A special emphasis is laid on luminescence dating with quartz, which is known for
systematic errors due to CW-OSL signal components with unequal physical behaviour.
This package enables an automatised workflow for signal component deconvolution
in combination with the R package 'Luminescence'.
}
\keyword{package}
