% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OSLcurve.R
\name{plot_OSLcurve}
\alias{plot_OSLcurve}
\title{Advanced plot function for decomposed OSL curves and component data}
\usage{
plot_OSLcurve(
  curve = NULL,
  components = NULL,
  display = "lin-log",
  zoom = 1,
  algorithm = "",
  title = NULL
)
}
\arguments{
\item{title}{}
}
\description{
Advanced plot function for decomposed OSL curves and component data
}
\section{Changelog}{

* 2019-03-06, DM: First reasonable version
* 2019-04-03, DM: Rebuild whole function
* 2019-05-03, DM: Checks now curve format; added x-axis zoom
* 2019-10-02, DM: Added background support and some little tweaks
}

\section{ToDo}{

* Show fitting formula
* Add pseudoLM-OSL presentation
* Cut data set while zooming to improve performance
* Debug residual curve zoom
* Get rid of library(ggpubr) to decrease dependencies
* Add early/late background substraction case
* display Residual square sum
}

\section{Last changed. 2019-11-25}{
NA
}

\examples{
test.components <- data.frame(name = c("fast","medium","slow"), lambda = c(1.5,0.5,0.1), n = c(1000,1000,10000))
test.curve <- simulate_OSLcurve(test.components, simulate.curve = TRUE, add.poisson.noise = TRUE, add.background = 20)
plot_OSLcurve(test.curve, test.components, display = "detailed")

}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
