% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OSLcurve.R
\name{plot_OSLcurve}
\alias{plot_OSLcurve}
\title{Advanced plot function for decomposed OSL curves and component data}
\usage{
plot_OSLcurve(
  curve = NULL,
  components,
  display = "detailed",
  zoom = 1,
  algorithm = "",
  hide_plot = FALSE,
  title = NULL
)
}
\arguments{
\item{display}{Choices: "detailed" (default), "components", "raw" ,"compare_lin", "compare_log", "lin-log"}

\item{hide_plot}{If true, plot is not drawn but can be catched by 'A <- plot_OSLcurve(...)'
and displayed later by grid.arrange(A)}

\item{title}{}
}
\description{
Advanced plot function for decomposed OSL curves and component data
}
\section{Changelog}{

* 2019-03-06, DM: First reasonable version
* 2019-04-03, DM: Rebuild whole function
* 2019-05-03, DM: Checks now curve format; added x-axis zoom
* 2019-10-02, DM: Added background component support and some little tweaks
* 2020-04-22, DM: Enabled hidden output
}

\section{ToDo}{

* REFACTORIZE CODE
* ! Add pseudoLM-OSL presentation !
* ! Put the ggplot building in its own sub-function !
* When drawing components without curve, skip residual curve
* Display fitting formula
* Display residual square sum
* Cut data set while zooming to improve performance
* Debug residual curve zoom
* Get rid of libraries 'scale' and 'ggpubr' to decrease dependencies
}

\section{Last changed. 2020-04-27}{
NA
}

\examples{
test.components <- data.frame(name = c("fast","medium","slow"), lambda = c(1.5,0.5,0.1), n = c(1000,1000,10000))
test.curve <- simulate_OSLcurve(test.components, simulate.curve = TRUE, add.poisson.noise = TRUE, add.background = 20)
plot_OSLcurve(test.curve, test.components, display = "detailed")

}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
