% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_OSLcurve.R
\name{plot_OSLcurve}
\alias{plot_OSLcurve}
\title{Advanced plot function for component resolved CW-OSL curves}
\usage{
plot_OSLcurve(
  curve = NULL,
  components,
  display = "detailed",
  show.legend = FALSE,
  show.intervals = FALSE,
  theme.set = ggplot2::theme_classic(),
  title = NULL,
  hide.plot = FALSE,
  filename = NULL
)
}
\arguments{
\item{curve}{\link{data.frame} or \link{matrix} or \linkS4class{RLum.Data.Curve} (\emph{optional}):
CW-OSL curve x-Axis: \verb{$time} or first column as measurement time (must have constant time intervals);
y-Axis: \verb{$signal} or second column as luminescence signal.
Other columns will be plotted as component curves, in case no argument \code{components} is defined.
If no input is given, a CW-OSL curve will be simulated with the parameters of \code{components}}

\item{components}{\link{data.frame} (\emph{optional}):
Table with OSL component parameters. Overrides component curves provides by \code{curve}.
Need to have at least the columns: \verb{$names}, \verb{$lambda} and \verb{$n}}

\item{display}{\link{character} (\emph{with default}):
Arrangement of graphs, see section \strong{Details}}

\item{show.legend}{\link{logical} (\emph{with default}):
Draws a legend in the top right corner of the first graph}

\item{show.intervals}{\link{logical} (\emph{with default}):
Draws vertical lines into the residual plot showing the signal bin intervals for Step 2 CW-OSL
decomposition (if available)}

\item{theme.set}{\link{ggplot2} object (\emph{with default}):
Graphical theme of the output plot. Input is forwarded to \link[ggplot2:theme_get]{ggplot2::theme_set}.
Recommended themes are \code{ggplot2::theme_minimal()}, \code{ggplot2::theme_classic()} and \code{ggplot2::theme_bw()},
see \link[ggplot2:ggtheme]{ggplot2::theme_bw} or \href{https://ggplot2.tidyverse.org/reference/ggtheme.html}{here} for
a full list}

\item{title}{\link{character} (\emph{with default}):
Plot title. Set \code{title = NULL} for no title}

\item{hide.plot}{\link{logical} (\emph{with default}):
If true, plot is not drawn but can still be saved as files or catched by \code{A <- plot_OSLcurve()}.
If catched, the plot can be drawn manually for example by using \link[gridExtra:arrangeGrob]{gridExtra::grid.arrange}}

\item{filename}{\link{character} (\emph{optional}):
File name or path to save the plot as image. If just a name is given, the image is
saved in the working directory. The image type is chosen by the file ending.
Allowed are \code{.pdf}, \code{.eps}, \code{.svg} (vector graphics), \code{.jpg}, \code{.png}, \code{.bmp} (pixel graphics)
and more, see \link[ggplot2:ggsave]{ggplot2::ggsave}}
}
\description{
Function for plotting CW-OSL curves and its signal components. It can handle data returned
by \link{fit_OSLcurve} or \link{decompose_OSLcurve}. Besides CW-OSL curves, pseudoLM-OSL and
residual are plotted.
}
\details{
\strong{Change graph types with parameter:} \code{display}

\tabular{ll}{
\code{"detailed"} \tab (default) Output plot consists of: Linear CW-OSL g, pseudoLM-OSL plot, Residual curve and component table \cr
\code{"lin"} \tab Linear CW-OSL plot only \cr
\code{"compare_lin"} \tab Linear CW-OSL plot with residual curve below and component table on bottom. Useful if two CW-OSL measurements shall be compared side by side \cr
\code{"log"} \tab CW-OSL plot with logarithmic y-Axis and linear x-Axis \cr
\code{"compare_log"} \tab CW-OSL plot with logarithmic y-Axis with residual curve below and component table on bottom. Useful if two CW-OSL measurements shall be compared side by side \cr
\code{"loglog"} \tab Double-logarithic CW-OSL plot \cr
\code{"LM"} \tab PseudoLM-OSL plot \cr
\code{"res"} \tab Plot of residual curve: Measurement minus Fitting model \cr
\code{"tab"} \tab Table of component parameters as image
}

PseudoLM-OSL curves are created using the transformation described by Bulur (2000).
The stimulation ramp duration is chosen double the CW-OSL duration.
See also Bos and Wallinga (2012) for a detailed explanation and discussion
}
\section{Last updates}{


2020-11-03, DM: Refactored code; Changed parameter \code{display} choices and added parameters \code{theme.set} and \code{show.legend}
}

\examples{

# Set some arbitary decay parameter for a dim CW-OSL measurement of quartz
components <- data.frame(name = c("fast", "medium", "slow"),
                         lambda = c(2, 0.5, 0.02),
                         n = c(1000, 1000, 10000))

# Simulate a CW-OSL curve including some signal noise
curve <- simulate_OSLcomponents(components, simulate.curve = TRUE, add.poisson.noise = TRUE)

# Display the simulated curve
plot_OSLcurve(curve, components)

}
\references{
Bos, A. J. J. and Wallinga, J.: How to visualize quartz OSL signal components,
Radiation Measurements, 47(9)

Bulur, E.: A simple transformation for converting CW-OSL curves to LM-OSL curves,
Radiation Measurements, 32(2)
}
\seealso{
\link{fit_OSLcurve}, [RLum.OSL_decomposition, \link{RLum.OSL_global_fitting}, \link{simulate_OSLcomponents}
}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}

Please cite the package the following way:

Mittelstraß, D., Schmidt, C., Beyer, J., Heitmann, J. and Straessner, A.:
Automated identification and separation of quartz CW-OSL signal components with R, \emph{in preparation}.
}
