% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_PhotoCrosssections.R
\name{plot_PhotoCrosssections}
\alias{plot_PhotoCrosssections}
\title{Plot comparison of the global OSL curve photo-crosssections with literature values}
\usage{
plot_PhotoCrosssections(
  fit.list,
  stimulation.intensity,
  stimulation.wavelength,
  title = NULL
)
}
\arguments{
\item{fit.list}{}
}
\value{

}
\description{
Plot comparison of the global OSL curve photo-crosssections with literature values
}
\section{Changelog}{

* 2019-10-08 Seperated from fit_OSLcurve()
* 2020-04-09 Changed from decay rates to photoionisation cross-sections; cleaned code
}

\section{ToDo}{

* Documentation
* Sometimes the red rectangle is not drawn (fit_OSLcurve(Batagai, K.max = 6)). Why?
* add literature values for other than just 470nm values
* remove library(XXXX)
}

\section{Last changed. 2020-04-09}{
NA
}

\examples{

}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
