% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.OSL_correction.R
\name{RLum.OSL_correction}
\alias{RLum.OSL_correction}
\title{Various tools for checking and manipulation CW-OSL curves in RLum.Analysis data sets}
\usage{
RLum.OSL_correction(
  object,
  record_type = "OSL",
  background_sequence = NULL,
  subtract_offset = 0,
  check_consistency = TRUE,
  check_signal_level = FALSE,
  tailor_records = FALSE,
  record_cut = 20,
  correct_PMTsaturation = NA,
  verbose = TRUE,
  report = FALSE
)
}
\arguments{
\item{object}{[RLum.Analysis-class] (**required**):
input object containing data for analysis, alternatively a [list] of
[RLum.Analysis-class] objects can be provided.}

\item{check_signal_level}{**Inactive**}

\item{tailor_records}{**Inactive**}

\item{correct_PMTsaturation}{**Inactive**}

\item{report}{}
}
\value{

}
\description{
Various tools for checking and manipulation CW-OSL curves in RLum.Analysis data sets
}
