% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.OSL_decomposition.R
\name{RLum.OSL_decomposition}
\alias{RLum.OSL_decomposition}
\title{Decomposes CW-OSL curves in RLum.Analysis data sets into its signal components}
\usage{
RLum.OSL_decomposition(
  object,
  K = NA,
  record_type = "OSL",
  decay_rates = NULL,
  error_calculation = "empiric",
  report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{object}{Object returned by [RLum.OSL_global_fitting()]}

\item{K}{Number of components. If not defined, 'K.selected' from the object returned by [RLum.OSL_global_fitting()] is used}

\item{decay_rates}{(optional) user-defined component decay rates}

\item{verbose}{}
}
\value{

}
\description{
Decomposes CW-OSL curves in RLum.Analysis data sets into its signal components
}
