% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RLum.OSL_global_fitting.R
\name{RLum.OSL_global_fitting}
\alias{RLum.OSL_global_fitting}
\title{Identifies CW-OSL signal components in RLum.Analysis data sets}
\usage{
RLum.OSL_global_fitting(
  object,
  max_components = 3,
  record_type = "OSL",
  F_threshold = 150,
  stimulation_intensity = 35,
  stimulation_wavelength = 470,
  report = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{verbose}{}
}
\value{

}
\description{
Identifies CW-OSL signal components in RLum.Analysis data sets
}
