% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_OSLcurves.R
\name{sum_OSLcurves}
\alias{sum_OSLcurves}
\title{Combine RLum OSL records to a global mean curve}
\usage{
sum_OSLcurves(
  object,
  record_type = "OSL",
  aliquot_selection = NULL,
  offset_value = 0,
  output.plot = TRUE,
  plot.first = TRUE,
  plot.global = TRUE,
  verbose = TRUE,
  title = "default",
  return.plot = FALSE
)
}
\arguments{
\item{object}{[RLum.Analysis-class] or [list](RLum.Analysis) (**required**):
data set of one or multiple aliquots containing containing luminescence measurements}

\item{record_type}{[character] (*with default*):
record type ("OSL","TL","IRSL") which shall be summed up}

\item{aliquot_selection}{[c] (*optional*):
vector specifying which items (aliquots) of a list of [RLum.Analysis-class] objects shall be combined}

\item{offset_value}{[numeric] (*with default*):
signal offset (background) which will be substracted from each record before combining}

\item{output.plot}{[logical] (*with default*):
returns a plot of the created superposition curve}

\item{return.plot}{}
}
\value{
A [data.frame] is returned, containing the created mean curve
}
\description{
The function adds up RLum records of one specific type
it is useful to create background records and enable component fitting
at sets of records with low signal-to-noise ratio
}
\section{Changelog}{

* 2018-05-23, DM: first version
* 2019-03-15, DM: rewritten for ggplots and new data format, renamend into sum_OSLcurves
}

\section{ToDo}{

* add more options for record selection (e.g. dose)
* interpolate x-axis if it doesn't match
* accept data.frames as input objects
* test if first value is zero
* add legend to plot
* add info box with number of OSL curves to plot
}

\section{Last changed}{
 2020-04-20
}

\examples{

}
\references{
Mittelstraß, D., Schmidt, C., Beyer, J., Heitmann, J. and Straessner, A.:
Automated identification and separation of quartz CW-OSL signal components with R, *in preparation*.
}
\seealso{
[fit_OSLcurve], [RLum.OSL_correction], [RLum.OSL_global_fitting]
}
\author{
Dirk Mittelstrass, \email{dirk.mittelstrass@luminescence.de}
}
