% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_OSLcurves.R
\name{sum_OSLcurves}
\alias{sum_OSLcurves}
\title{Combine RLum OSL records to a global mean curve}
\usage{
sum_OSLcurves(
  object,
  record_type = "OSL",
  aliquot_selection = NULL,
  offset_value = 0,
  output.plot = TRUE,
  plot.first = TRUE,
  plot.global = TRUE,
  verbose = TRUE,
  title = "default",
  return.plot = FALSE
)
}
\arguments{
\item{object}{\linkS4class{RLum.Analysis} or \href{RLum.Analysis}{list} (\strong{required}):
data set of one or multiple aliquots containing containing luminescence measurements}

\item{record_type}{\link{character} (\emph{with default}):
record type ("OSL","TL","IRSL") which shall be summed up}

\item{aliquot_selection}{\link{c} (\emph{optional}):
vector specifying which items (aliquots) of a list of \linkS4class{RLum.Analysis} objects shall be combined}

\item{offset_value}{\link{numeric} (\emph{with default}):
signal offset (background) which will be substracted from each record before combining}

\item{output.plot}{\link{logical} (\emph{with default}):
returns a plot of the created superposition curve}
}
\value{
A \link{data.frame} is returned, containing the created mean curve
}
\description{
The function adds up RLum records of one specific type
it is useful to create background records and enable component fitting
at sets of records with low signal-to-noise ratio
}
\section{Changelog}{

\itemize{
\item 2018-05-23, DM: first version
\item 2019-03-15, DM: rewritten for ggplots and new data format, renamend into sum_OSLcurves
}
}

\section{ToDo}{

\itemize{
\item add more options for record selection (e.g. dose)
\item interpolate x-axis if it doesn't match
\item accept data.frames as input objects
\item test if first value is zero
\item add legend to plot
\item add info box with number of OSL curves to plot
}
}

\section{Last changed}{
 2020-04-20
}

\references{
Dirk Mittelstrass, Christoph Schmidt, Sebastian Kreutzer, ... .\emph{in preperation}. Algebraic CW-OSL
signal component decomposition of quartz and its use for dose evaluation within the R luminescence package
}
\seealso{
\link{calc_CWOSLcomponents}, \link{analyse_SAR.CWOSL.deconvolved}, \link{calc_deconvolution.intervals}
}
\author{
Dirk Mittelstrass, TU Dresden (Germany), \email{dirk.mittelstrass@luminescence.de}
}
